@annrpom 2023-11-28
Note: the amount of roundtrips increases by 1 with the new change, but this doesn't seem to affect the overall impact of this change (i.e. even with a setup of a single database, the speed of the benchmark query is improved by database-aware scans)